В DAO добавлены методы delete, getByID. Во всех можелях исправлены ID на int

+++ finalYearProject/src/main/java/com/work/dao/api/EmployeeDAO.java
@@ -3,5 +3,5 @@ package com.work.dao.api;
 import com.work.model.Employee;
 
 public interface EmployeeDAO extends DAO<Employee> {
-    Employee getEmployeeByID(long id);
+    Employee getEmployeeByID(int id);
 }
